argoslabs/data/docx2txt/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# ...
pypandoc